Update order route for delete, now on delete, the price of the order is returned simulating the price to pay

+++ src/main/java/ch/hearc/jee2024/project/IOC/OrderController.java
@@ -6,6 +6,8 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.List;
+
 
 @RestController
 @RequestMapping("/orders")
@@ -56,10 +58,15 @@ public class OrderController {
     }
 
 
-    @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteOrder(@PathVariable Long id) {
-        orderService.deleteOrder(id);
-        return ResponseEntity.noContent().build();
+    @DeleteMapping("/{id}/checkout")
+    public ResponseEntity<Double> deleteOrder(@PathVariable Long id) {
+        Order order = orderService.getOrderById(id).orElse(null);
+        if(order == null) {
+            return ResponseEntity.notFound().build();
+        }else {
+            orderService.deleteOrder(id);
+            return ResponseEntity.ok(order.getTotalPrice());
+        }
     }
 
 
@@ -73,5 +80,7 @@ public class OrderController {
         Page<Order> orders = orderService.getOrders(page, size, sortBy, direction);
         return ResponseEntity.ok(orders);
     }
+
+
 }
